+++ src/email/email_server.rb
@@ -65,6 +65,10 @@ OpenTelemetry.meter_provider.add_metric_reader(otlp_metric_exporter)
 meter = OpenTelemetry.meter_provider.meter("email")
 $confirmation_counter = meter.create_counter("demo.notification.confirmations", unit: "1", description: "Counts the number of order confirmation emails sent")
 
+get "/healthz" do
+  status 200
+end
+
 post "/send_order_confirmation" do
   data = JSON.parse(request.body.read, object_class: OpenStruct)