+++ src/frontend/pages/api/healthz.ts
@@ -1,0 +1,10 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+import type { NextApiRequest, NextApiResponse } from 'next';
+
+const handler = (_req: NextApiRequest, res: NextApiResponse) => {
+  return res.status(200).end();
+};
+
+export default handler;